@@ -21,10 +21,11 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 		<maven.compiler.encoding>UTF-8</maven.compiler.encoding>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
+		<maven.compiler.source>17</maven.compiler.source>
+		<maven.compiler.target>17</maven.compiler.target>
 		<maven.test.skip>true</maven.test.skip>
 		<!-- plugin -->
+		<maven-compiler-plugin.version>3.14.0</maven-compiler-plugin.version>
 		<maven-source-plugin.version>3.3.1</maven-source-plugin.version>
 		<maven-javadoc-plugin.version>3.11.2</maven-javadoc-plugin.version>
 		<maven-gpg-plugin.version>3.2.8</maven-gpg-plugin.version>
@@ -32,25 +33,20 @@
 
 		<!-- base -->
 		<slf4j-api.version>2.0.17</slf4j-api.version>
-		<junit-jupiter.version>5.13.3</junit-jupiter.version>
+		<junit-jupiter.version>5.13.4</junit-jupiter.version>
 
 		<!-- spring -->
-		<spring.version>5.3.39</spring.version>
-		<spring-boot.version>2.7.18</spring-boot.version>
-		<!-- database -->
-		<mybatis-spring-boot-starter.version>2.3.2</mybatis-spring-boot-starter.version>
-		<mysql-connector-j.version>9.3.0</mysql-connector-j.version>
+		<spring.version>6.2.10</spring.version>
+		<spring-boot.version>3.5.4</spring-boot.version>
+		<!-- mybatis & db -->
+		<mybatis-spring-boot-starter.version>3.0.5</mybatis-spring-boot-starter.version>
+		<mysql-connector-j.version>9.4.0</mysql-connector-j.version>
 
-		<!-- xxl-tool( + support) -->
-		<xxl-tool.version>1.5.0</xxl-tool.version>
-		<gson.version>2.13.1</gson.version>
-
+		<!-- xxl-tool -->
+		<xxl-tool.version>2.0.0</xxl-tool.version>
+		<!-- xxl-sso (+xxl-tool、gson) -->
+		<xxl-sso.version>2.0.0</xxl-sso.version>
 	</properties>
-
-	<build>
-		<plugins>
-		</plugins>
-	</build>
 
 	<licenses>
 		<license>
@@ -58,7 +54,6 @@
 			<url>https://opensource.org/licenses/GPL-3.0</url>
 		</license>
 	</licenses>
-
 	<scm>
 		<tag>master</tag>
 		<url>https://github.com/xuxueli/xxl-mq.git</url>
@@ -74,6 +69,11 @@
 		</developer>
 	</developers>
 
+	<build>
+		<plugins>
+		</plugins>
+	</build>
+
 	<profiles>
 
 		<profile>
@@ -86,6 +86,17 @@
 			</modules>
 			<build>
 				<plugins>
+					<!-- Compiler -->
+					<plugin>
+						<groupId>org.apache.maven.plugins</groupId>
+						<artifactId>maven-compiler-plugin</artifactId>
+						<version>${maven-compiler-plugin.version}</version>
+						<configuration>
+							<source>${maven.compiler.source}</source>
+							<target>${maven.compiler.target}</target>
+							<parameters>true</parameters>
+						</configuration>
+					</plugin>
 					<!-- Source -->
 					<plugin>
 						<groupId>org.apache.maven.plugins</groupId>
@@ -144,7 +155,6 @@
 							<publishingServerId>central</publishingServerId>
 							<excludeArtifacts>
 								<artifact>xxl-mq-admin</artifact>
-								<artifact>xxl-mq-samples</artifact>
 								<artifact>xxl-mq-samples</artifact>
 								<artifact>xxl-mq-samples-frameless</artifact>
 								<artifact>xxl-mq-samples-springboot</artifact>
