@@ -15,7 +15,7 @@ import com.xxl.tool.response.Response;
 import com.xxl.tool.response.ResponseCode;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
